experiments/4dbinfer/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pandas>=1.5.0
numpy>=1.21.0
pydantic>=1.10.0
pyyaml>=6.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0